requirements.txt
flask
paho-mqtt
opencv-python-headless
numpy
ultralytics